# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Komşu Coordination API",
    "description": "Human-led coordination development foundation; all AI outputs require review.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/retention/plans": {
      "get": {
        "summary": "Retention Plans",
        "operationId": "retention_plans_api_v1_retention_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Retention Plan",
        "operationId": "create_retention_plan_api_v1_retention_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionScheduleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/retention/plans/{plan_id}/approve": {
      "post": {
        "summary": "Approve Retention Plan",
        "operationId": "approve_retention_plan_api_v1_retention_plans__plan_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionApproveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/retention/plans/{plan_id}/cancel": {
      "post": {
        "summary": "Cancel Retention Plan",
        "operationId": "cancel_retention_plan_api_v1_retention_plans__plan_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionCancelIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/retention/holds": {
      "get": {
        "summary": "Legal Holds",
        "operationId": "legal_holds_api_v1_retention_holds_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Add Legal Hold",
        "operationId": "add_legal_hold_api_v1_retention_holds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalHoldIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/retention/holds/{hold_id}/release": {
      "post": {
        "summary": "Remove Legal Hold",
        "operationId": "remove_legal_hold_api_v1_retention_holds__hold_id__release_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hold_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Hold Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalHoldReleaseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports": {
      "post": {
        "summary": "Receive",
        "operationId": "receive_api_v1_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/{report_id}/audio": {
      "post": {
        "summary": "Receive Audio",
        "operationId": "receive_audio_api_v1_reports__report_id__audio_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioAccepted"
                }
              }
            }
          },
          "200": {
            "description": "Identical audio replay"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RetentionApproveIn": {
        "properties": {
          "expected_status": {
            "type": "string",
            "const": "SCHEDULED",
            "title": "Expected Status"
          },
          "acknowledge_irreversible_purge": {
            "type": "boolean",
            "const": true,
            "title": "Acknowledge Irreversible Purge"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_status",
          "acknowledge_irreversible_purge"
        ],
        "title": "RetentionApproveIn"
      },
      "RetentionCancelIn": {
        "properties": {
          "expected_status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "APPROVED"
            ],
            "title": "Expected Status"
          },
          "reason_code": {
            "type": "string",
            "enum": [
              "PLAN_REPLACED",
              "INCIDENT_REOPENED",
              "ENTERED_IN_ERROR"
            ],
            "title": "Reason Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_status",
          "reason_code"
        ],
        "title": "RetentionCancelIn"
      },
      "RetentionScheduleIn": {
        "properties": {
          "cutoff_at": {
            "type": "string",
            "format": "date-time",
            "title": "Cutoff At"
          },
          "execute_after": {
            "type": "string",
            "format": "date-time",
            "title": "Execute After"
          },
          "reason_code": {
            "type": "string",
            "enum": [
              "INCIDENT_ENDED",
              "RETENTION_EXPIRED",
              "DATA_SUBJECT_REQUEST"
            ],
            "title": "Reason Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cutoff_at",
          "execute_after",
          "reason_code"
        ],
        "title": "RetentionScheduleIn"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}